% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_aggregate_fp.R
\name{aggregate_fp}
\alias{aggregate_fp}
\title{weight_samples}
\usage{
aggregate_fp(division_level_data, population_data, posterior_samples)
}
\arguments{
\item{division_level_data}{data frame with country codes and corresponding aggregate level}

\item{population_data}{population data, union and years of the samples}

\item{posterior_samples}{posteriors samples from \code{\link{posterior_samples_array_format}}}
}
\description{
weight_samples
}
\examples{
dimnames(posterior_samples) <- list(division_numeric_code, NULL, NULL, NULL) #provide corresponding division numeric codes in posterior_samples attributes
population_data <- population_counts \%>\%
  dplyr::filter(is_in_union == union) \%>\%
  dplyr::filter(mid_year <= last_year) \%>\%
  dplyr::filter(mid_year >= first_year)
division_level_data <- divisions \%>\%
  mutate(division_level = region_numeric_code)\%>\%
  select(division_numeric_code, division_level)
posterior_samples_list <- weight_samples(division_level_data, population_data, posterior_samples)
}
