% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_samples.R
\name{aggregate_with_weights}
\alias{aggregate_with_weights}
\title{aggregate_with_weights}
\usage{
aggregate_with_weights(posterior_samples, weights)
}
\arguments{
\item{weights}{\emph{'Vector'} A vector of weights selected from \code{\link[weight_generator]{weight_generator}}}
}
\value{
\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions
}
\description{
aggregate_with_weights
}
