% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_calc_fp.R
\name{calc_fp}
\alias{calc_fp}
\title{calc_fp}
\usage{
calc_fp(posterior_samples, population_data, first_year)
}
\arguments{
\item{posterior_samples}{}

\item{population_data}{}

\item{first_year}{}
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
calc_fp
}
