% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraceptive_use_filter_impute.R
\name{contraceptive_use_filter_impute}
\alias{contraceptive_use_filter_impute}
\title{contraceptive_use_filter}
\usage{
contraceptive_use_filter_impute(
  survey_dataset,
  custom_data_indicator,
  is_in_union,
  division_numeric_code,
  subnational
)
}
\arguments{
\item{survey_dataset}{\emph{\sQuote{Data frame}} If this argument is not specified the UNPD dataset \code{\link{contraceptive_use}} is used.}

\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}

\item{subnational}{'\emph{\sQuote{Logical}} If TRUE runs the sub national model.}
}
\description{
contraceptive_use_filter
}
