% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_data.R
\name{core_data}
\alias{core_data}
\title{core_data}
\usage{
core_data(
  is_in_union,
  surveydata_filepath,
  division_numeric_code,
  first_year,
  last_year,
  subnational
)
}
\description{
Creates core_data that consists of manipulated observations and run settings for \code{\link{do_1country_run}}
}
