% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fit_fp_c.R
\name{fit_fp_c}
\alias{fit_fp_c}
\title{Fit one country model}
\usage{
fit_fp_c(is_in_union, ...)
}
\arguments{
\item{...}{Please see \code{\link{fit_fp_csub}} for additional arguments}
}
\description{
A wrapper for \code{\link{fit_fp_csub}} capable of running multiple unions.
}
