% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fit_fp_c.R
\name{fit_fp_c_autosave}
\alias{fit_fp_c_autosave}
\title{Fit one country model and automatically save output}
\usage{
fit_fp_c_autosave(runname = NULL, ...)
}
\arguments{
\item{runname}{\emph{\sQuote{Character}} The name given to the file generated by this wrapper.}

\item{...}{Please see \code{\link{fit_fp_csub}} for additional arguments.}
}
\description{
A wrapper for \code{\link{fpem_one_country}} which saves the output. The output can be automatically read in by proceeding fpem functions with the runname argument specified.
}
