% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fp_c.R
\name{fit_fp_csub}
\alias{fit_fp_csub}
\title{Fit one country on union model}
\usage{
fit_fp_csub(
  survey_dataset,
  division_numeric_code,
  is_in_union,
  first_year,
  last_year,
  subnational,
  service_stats_dataset,
  diagnostic,
  params_to_save,
  nchains,
  niter,
  nburnin,
  nthin
)
}
\arguments{
\item{survey_dataset}{\emph{\sQuote{Data frame}} If this argument is not specified the UNPD dataset \code{\link{contraceptive_use}} is used.}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}

\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}

\item{last_year}{\emph{\sQuote{Numeric}} The last year of model estimates in output. The model will be fit to all data, including dates after this date if available.}

\item{subnational}{'\emph{\sQuote{Logical}} If TRUE runs the sub national model.}

\item{service_stats_dataset}{\emph{\sQuote{Data frame}}}

\item{diagnostic}{'\emph{\sQuote{Logical}} If TRUE saves the full jags output.}
}
\description{
The central subroutine for model fitting. Fits a single country model for in-union or not-in-union women.
}
