% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_one_country.R
\name{fpem_1country_1union}
\alias{fpem_1country_1union}
\title{Run one country model}
\usage{
fpem_1country_1union(
  is_in_union = "Y",
  surveydata_filepath = NULL,
  service_stats = FALSE,
  service_stats_filepath = NULL,
  division_numeric_code,
  first_year = NULL,
  last_year,
  subnational = FALSE,
  diagnostic = FALSE,
  params_to_save = c("mod.ct", "unmet.ct", "trad.ct", "mu.in",
    "logitratio.yunmet.hat.i"),
  nchains = 3,
  niter = 2500,
  nburnin = 500,
  nthin = max(1, floor((niter - nburnin)/1000))
)
}
\arguments{
\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}

\item{surveydata_filepath}{\emph{\sQuote{Character}} Path to survey data. Survey data should be a .csv with the following format \code{\link{contraceptive_use}}.}

\item{service_stats}{\emph{\sQuote{Logical}} If TRUE and if a service stat file path is provided, the model is run using service statistic data. Default Service stats are not public. Requires private repository FPcounts/fpemservicestat package.}

\item{service_stats_filepath}{\emph{\sQuote{Character}}}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}

\item{last_year}{\emph{\sQuote{Numeric}} The last year of model estimates in output. The model will be fit to all data, including dates after this date if available.}

\item{subnational}{'\emph{\sQuote{Logical}} If TRUE runs the sub national model.}

\item{diagnostic}{'\emph{\sQuote{Logical}} If TRUE saves the full jags output.}
}
\value{
\emph{\sQuote{List}}
\enumerate{
  \item \strong{posterior_samples}  \emph{\sQuote{Numeric array}} An array of samples of dimension chains x samples x years x proportions.
  \item \strong{core_data}          \emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link{core_data}}.
}
}
\description{
Runs the family planning estimation model and returns samples and bias adjusted observations for respective run.
}
\examples{
See the reposiotry url in references for detailed examples
}
\references{
\url{https://github.com/FPcounts/FPEM}
}
