% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet_calculate.R,
%   R/fpet_calculate_indicators.R, R/fpet_calculate_results.R
\name{fpem_calculate_results}
\alias{fpem_calculate_results}
\title{Calculate results from samples}
\usage{
fpem_calculate_results(
  posterior_samples,
  country_population_counts,
  first_year
)

fpem_calculate_results(
  posterior_samples,
  country_population_counts,
  first_year
)

fpem_calculate_results(
  posterior_samples,
  country_population_counts,
  first_year
)
}
\arguments{
\item{country_population_counts}{\emph{'Numeric Vector'} A vector of population counts selected from \code{\link[get_population_counts]{get_population_counts}}}

\item{runlist}{\emph{'list'} a list with core_data and posterior samples}
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.

\emph{'Data.frame'} A data.frame of point estimates in long format.

\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
Returns point estimates from posterior samples in long format.

Returns point estimates from posterior samples in long format.

Returns point estimates from posterior samples in long format.
}
