% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_get_plots.R
\name{fpem_get_plots.all_union}
\alias{fpem_get_plots.all_union}
\title{plot country results}
\usage{
\method{fpem_get_plots}{all_union}(runlist, results, indicators)
}
\arguments{
\item{results}{\emph{'Data.frame'} Results data from \code{\link{fpem_calculate_results}}}

\item{indicators}{name of indicators from results to be plotted}
}
\value{
list of plots
}
\description{
plot country results
}
