% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_get_plots.R
\name{fpem_get_plots.single_union}
\alias{fpem_get_plots.single_union}
\title{plot country results}
\usage{
\method{fpem_get_plots}{single_union}(runlist, results, indicators, compare_to_global)
}
\arguments{
\item{results}{\emph{'Data.frame'} Results data from \code{\link{fpem_calculate_results}}}

\item{indicators}{name of indicators from results to be plotted}

\item{compare_to_global}{logical, if TRUE plots estimates from global model with dotted lines}
}
\value{
list of plots
}
\description{
plot country results
}
