% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_get_plots.R
\name{fpem_get_plots_autosave}
\alias{fpem_get_plots_autosave}
\title{plot country results}
\usage{
fpem_get_plots_autosave(runname, indicators, compare_to_global)
}
\arguments{
\item{indicators}{name of indicators from results to be plotted}

\item{compare_to_global}{logical, if TRUE plots estimates from global model with dotted lines}
}
\value{
list of plots
}
\description{
plot country results
}
