% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_get_results.R
\name{fpem_get_results}
\alias{fpem_get_results}
\title{Get results from samples}
\usage{
fpem_get_results(...)
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
Returns point estimates from posterior samples in long format.
}
