% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_one_country.R
\name{fpem_one_country}
\alias{fpem_one_country}
\title{Run one country model}
\usage{
fpem_one_country(is_in_union, ...)
}
\arguments{
\item{...}{Please see \code{\link{fpem_1country_1union}} for additional arguments}
}
\description{
A wrapper for \code{\link{fpem_1country_1union}} capable of running multiple unions.
}
