% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_one_country.R
\name{fpem_one_country_autosave}
\alias{fpem_one_country_autosave}
\title{Run one country model and automatically save output}
\usage{
fpem_one_country_autosave(runname = NULL, ...)
}
\arguments{
\item{runname}{\emph{\sQuote{Character}} The name given to the file generated by this wrapper.}

\item{...}{Please see \code{\link{fpem_1country_1union}} for additional arguments.}
}
\description{
A wrapper for \code{\link{fpem_one_country}} which saves the output. The output can be automatically read in by proceeding fpem functions with the runname argument specified.
}
