% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fpem_plot_country_results}
\alias{fpem_plot_country_results}
\title{plot country results}
\usage{
fpem_plot_country_results(country_results, core_data, indicators)
}
\arguments{
\item{country_results}{\emph{'Data.frame'} Results data from \code{\link{fpem_calculate_results}}}

\item{core_data}{data list from \code{\link{core_data}}}

\item{indicators}{name of indicators from results to be plotted}
}
\value{
list of plots
}
\description{
plot country results
}
