% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_results.R
\name{fpem_results}
\alias{fpem_results}
\title{Get results from samples}
\usage{
fpem_results(runlist, country_population_counts = NULL)
}
\arguments{
\item{runlist}{\emph{'list'} a list with core_dat and posteior samples}

\item{country_population_counts}{\emph{'Numeric Vector'} A vector of population counts selected from \code{\link[get_population_counts]{get_population_counts}}}
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
Maps multiple sets of runs to \code{\link{fpem_calculate_results}}
}
