% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_results.R
\name{fpem_results_autosave}
\alias{fpem_results_autosave}
\title{Get results from samples}
\usage{
fpem_results_autosave(runname = NULL, country_population_counts = NULL)
}
\arguments{
\item{runname}{}

\item{country_population_counts}{}
}
\value{

}
\description{
A wrapper for \code{\link{fpem_results}} which reads in the model fit from the respective `runname`. The output is saved in the results directory with the same `runname`. The output can be automatically read in by proceeding fpem functions with the `runname` argument specified.
}
