% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet_results.R
\name{fpet_calculate_indicators}
\alias{fpet_calculate_indicators}
\title{Get results from samples}
\usage{
fpet_calculate_indicators(runlist, country_population_counts = NULL)
}
\description{
Maps multiple sets of runs to \code{\link{fpem_calculate_results}}
}
