% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet_results.R
\name{fpet_calculate_results}
\alias{fpet_calculate_results}
\title{Calculate results from samples}
\usage{
fpet_calculate_results(fit, country_population_counts = NULL)
}
\arguments{
\item{country_population_counts}{\emph{'Numeric Vector'} A vector of population counts selected from \code{\link[get_population_counts]{get_population_counts}}}

\item{runlist}{\emph{'list'} a list with core_data and posterior samples}
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
Returns point estimates from posterior samples in long format.
}
