% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet_one_country.R
\name{fpet_fit_model}
\alias{fpet_fit_model}
\title{Run one country model}
\usage{
fpet_fit_model(is_in_union, ...)
}
\arguments{
\item{...}{Please see \code{\link{fpet_fit_1country_1union_model}} for additional arguments}
}
\description{
A wrapper for \code{\link{fpet_fit_1country_1union_model}} capable of running multiple unions.
}
