% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet_one_country.R
\name{fpet_fit_model_autosave}
\alias{fpet_fit_model_autosave}
\title{Run one country model and automatically save output}
\usage{
fpet_fit_model_autosave(runname = NULL, ...)
}
\arguments{
\item{runname}{\emph{\sQuote{Character}} The name given to the file generated by this wrapper.}

\item{...}{Please see \code{\link{fpet_fit_1country_1union_model}} for additional arguments.}
}
\description{
A wrapper for \code{\link{fpem_one_country}} which saves the output. The output can be automatically read in by proceeding fpem functions with the runname argument specified.
}
