% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet.R
\name{get_contraceptive_use_any}
\alias{get_contraceptive_use_any}
\title{Get total CPR}
\usage{
get_contraceptive_use_any(posterior_samples, first_year)
}
\arguments{
\item{posterior_samples}{\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions}

\item{first_year}{`integer` Earliest year represented in the data}
}
\value{
`data.frame` Values by year and percentile
}
\description{
`total_cpr = modern_cpr + traditional_cpr`
}
