% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_contraceptive_use_traditional}
\alias{get_contraceptive_use_traditional}
\title{Get traditional CPR}
\usage{
get_contraceptive_use_traditional(posterior_samples, first_year)
}
\arguments{
\item{posterior_samples}{}

\item{first_year}{`integer` Earliest year represented in the data}
}
\value{
`data.frame` Values by year and percentile
}
\description{
Get traditional CPR
}
