% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet.R
\name{get_non_use}
\alias{get_non_use}
\title{Get non-use}
\usage{
get_non_use(posterior_samples, first_year)
}
\arguments{
\item{posterior_samples}{\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions}

\item{first_year}{`integer` Earliest year represented in the data}
}
\value{
`data.frame` Values by year and percentile
}
\description{
`non_use = 1 - total_cpr`
}
