% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_posterior_samples_for_all_indicators}
\alias{get_posterior_samples_for_all_indicators}
\title{Get posterior samples for all indicators}
\usage{
get_posterior_samples_for_all_indicators(posterior_samples, years)
}
\arguments{
\item{posterior_samples}{\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions}

\item{first_year}{`integer` Earliest year represented in the data}
}
\value{
`data.frame` Values by indicator and year
}
\description{
Get posterior samples for all indicators
}
