% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet.R
\name{get_proportions}
\alias{get_proportions}
\title{Get proportions}
\usage{
get_proportions(posterior_samples, first_year, transformer)
}
\arguments{
\item{posterior_samples}{\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions}

\item{first_year}{`integer` Earliest year represented in the data}

\item{transformer}{`function` Computes the desired result}
}
\value{
`data.frame` Values by year and percentile
}
\description{
Get proportions
}
