% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemreporting.R
\name{get_unmet_need_any}
\alias{get_unmet_need_any}
\title{Get unmet need}
\usage{
get_unmet_need_any(posterior_samples, first_year)
}
\arguments{
\item{posterior_samples}{\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions}

\item{first_year}{`integer` Earliest year represented in the data}
}
\value{
`data.frame` Values by year and percentile
}
\description{
Get unmet need
}
