% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_se.R
\name{impute_se}
\alias{impute_se}
\title{Impute se}
\usage{
impute_se(wide_data)
}
\arguments{
\item{wide_data}{\emph{'Data.frame'} A data.frame from \code{\link[contraceptive_use]{contraceptive_use}}}
}
\value{
\emph{'Data.frame'} The input data with se imputed
}
\description{
Imputes missing se based on Cahill et al 2017
}
