% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service_stats.R
\name{list_service_stats}
\alias{list_service_stats}
\title{list_service_stats}
\usage{
list_service_stats(
  service_stats = FALSE,
  service_stats_filepath,
  model_seq_years,
  division_numeric_code,
  first_year_observed = NULL
)
}
\arguments{
\item{service_stats}{\emph{\sQuote{Logical}} If TRUE and if a service stat file path is provided, the model is run using service statistic data. Default Service stats are not public.}

\item{service_stats_filepath}{File path to service stat csv}

\item{model_seq_years}{Years of model estimation from\code{\link{year_sequence_maker}}.}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}
}
\description{
fit_fp_cservice statistics data
}
