% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot_fp_c.R
\name{plot_fp_c}
\alias{plot_fp_c}
\title{plot country results}
\usage{
plot_fp_c(runlist, results, indicators, ...)
}
\arguments{
\item{results}{\emph{'Data.frame'} Results data from \code{\link{fpem_calculate_results}}}

\item{indicators}{name of indicators from results to be plotted}
}
\value{
list of plots
}
\description{
plot country results
}
