% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples_all_women}
\alias{posterior_samples_all_women}
\title{Get posterior samples for all women}
\usage{
posterior_samples_all_women(
  in_union_posterior_samples,
  not_in_union_posterior_samples,
  core_data
)
}
\arguments{
\item{in_union_posterior_samples}{\emph{`array`} An array of n chains x n iterations x n years x n proportions}

\item{not_in_union_posterior_samples}{\emph{`array`} An array of n chains x n iterations x n years x n proportions}

\item{core_data}{}
}
\value{
\emph{`array`} Posterior samples for all women
}
\description{
Combine posterior samples from in-union and not-in-union to obtain samples for all women
}
