% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_plots.R
\name{process_global}
\alias{process_global}
\title{Reformat UNPD global data to long format}
\usage{
process_global(data, code)
}
\arguments{
\item{data}{emph{'Data.frame'} Estimates for modern/trad/unmet obtained from posterior of global model in wide format}

\item{code}{emph{'Numeric'} A country code to filter on since we only run one country at a time on our end}
}
\value{
emph{'Data.frame} A long format data of specific country to be plotted
}
\description{
Reformat UNPD global data to long format
}
\examples{
get_global_estimates_married() \%>\% filter(par =="unmet") \%>\% process_global(code)
}
