% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documents.R
\docType{data}
\name{service_stats}
\alias{service_stats}
\title{Service statistics}
\format{
A data frame with 6 variables:
\describe{
  \item{"division_numeric_code"}{country code}
  \item{"name"}{country name}
  \item{"year"}{mid year of observation period}
  \item{"emu"}
  \item{"ss_type"}
  ...
}
}
\source{

}
\usage{
service_stats
}
\description{
Service stat data must contain the following variables. Names must be as documented (eg. all lowercase). `fpemdata::service_stats` returns a vector of column names. Data not publicly available.
}
\keyword{datasets}
