% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_check.R
\name{validate_this}
\alias{validate_this}
\title{Validate this}
\usage{
validate_this(format_list, data)
}
\arguments{
\item{format_list}{}

\item{data}{}
}
\value{
an informative error message if data does not pass validation
}
\description{
Validate this
}
