% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_aggregate_fp.R
\name{weight_division_match}
\alias{weight_division_match}
\title{weight_division_match}
\usage{
weight_division_match(division_level_data, population_data, posterior_samples)
}
\value{
posterior_samples array with divisions in the order to match weights which will be calculated
}
\description{
weight_division_match
}
