% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_and_aggregate.R
\name{weight_generator}
\alias{weight_generator}
\title{weight_generator}
\usage{
weight_generator(division_level_data, population_data)
}
\arguments{
\item{division_level_data}{data frame with country codes and corresponding aggregate level}

\item{population_data}{population data, union and years of the samples}
}
\value{
data.frame
}
\description{
weight_generator
}
