% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_and_aggregate.R
\name{weight_input_checker}
\alias{weight_input_checker}
\title{weight_input_checker}
\usage{
weight_input_checker(division_level_data, population_data, posterior_samples)
}
\arguments{
\item{division_level_data}{data frame with country codes and corresponding aggregate level}

\item{population_data}{population data, union and years of the samples}

\item{posterior_samples}{posteriors samples from \code{\link{posterior_samples_array_format}}}
}
\value{
informative error messages when inputs are incorrect
}
\description{
weight_input_checker
}
