% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_and_aggregate.R
\name{weight_multiplier}
\alias{weight_multiplier}
\title{weight_multiplier}
\usage{
weight_multiplier(posterior_samples, weights)
}
\arguments{
\item{posterior_samples}{posteriors samples from \code{\link{posterior_samples_array_format}}}

\item{weights}{\emph{'Vector'} A vector of weights selected from \code{\link[weight_generator]{weight_generator}}}
}
\value{
\emph{'Numeric array'} An array of samples of dimension chains x samples x years x proportions
}
\description{
weight_multiplier
}
