% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_calc_fp_c.R
\name{calc_fp_c}
\alias{calc_fp_c}
\title{Calculate fp indicators from samples}
\usage{
calc_fp_c(fit, country_population_counts = NULL)
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
Maps multiple sets of runs to \code{\link{calc_fp}}
}
