% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fp_c.R
\name{calc_fp_c}
\alias{calc_fp_c}
\title{calculate fp estimates for one country}
\usage{
calc_fp_c(fit, population_data = population_counts)
}
\arguments{
\item{fit}{\emph{\sQuote{List}} The fit object from \code{\link{fit_fp_c}}}

\item{population_data}{\emph{\sQuote{Data frame}} Population count data in the format of \code{\link{population_counts}}}
}
\value{

}
\description{
calculate fp estimates for one country
}
