% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fp_c.R
\name{fit_fp_c}
\alias{fit_fp_c}
\title{Fit one country model}
\usage{
fit_fp_c(
  survey_dataset = imputed_data$contraceptive_use_imputed,
  division_numeric_code = NULL,
  is_in_union = "Y",
  first_year = NULL,
  last_year = NULL,
  subnational = FALSE,
  service_stats_dataset = NULL,
  diagnostic = FALSE,
  params_to_save = c("mod.ct", "unmet.ct", "trad.ct", "mu.in",
    "logitratio.yunmet.hat.i"),
  nchains = 10,
  niter = 2500,
  nburnin = 500,
  nthin = max(1, floor((niter - nburnin) * nchains/2000))
)
}
\arguments{
\item{survey_dataset}{\emph{\sQuote{Data frame}} If this argument is not specified the UNPD dataset \code{\link{contraceptive_use}} is used.}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}

\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}

\item{last_year}{\emph{\sQuote{Numeric}} The last year of model estimates in output. The model will be fit to all data, including dates after this date if available.}

\item{subnational}{'\emph{\sQuote{Logical}} If TRUE runs the sub national model.}

\item{service_stats_dataset}{\emph{\sQuote{Data frame}}}

\item{diagnostic}{'\emph{\sQuote{Logical}} If TRUE saves the full jags output.}
}
\value{
\emph{\sQuote{List}}
\enumerate{
  \item \strong{posterior_samples}  \emph{\sQuote{Numeric array}} An array of samples of dimension chains x samples x years x proportions.
  \item \strong{core_data}          \emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link{core_data}}.
}
}
\description{
Fits the family planning estimation model and returns samples and bias adjusted observations for respective run.
}
\examples{
See the reposiotry url in references for detailed examples
}
\references{
\url{https://github.com/FPcounts/FPEM}
}
