% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_one_country.R
\name{fpem_one_country}
\alias{fpem_one_country}
\title{Run one country model}
\usage{
fpem_one_country(is_in_union, ...)
}
\arguments{
\item{is_in_union}{\emph{\sQuote{Character}} "Y" if women are in union.}
}
\value{
\emph{\sQuote{List}}
\enumerate{
  \item \strong{posterior_samples}  \emph{\sQuote{Numeric array}} An array of samples of dimension chains x samples x years x proportions.
  \item \strong{core_data}          \emph{\sQuote{Data.frame}} The processed data associated with the model run from \code{\link{core_data}}.
}
}
\description{
Runs the family planning estimation model and returns samples and bias adjusted observations for respective run.
}
\examples{
See the reposiotry url in references for detailed examples
}
\references{
\url{https://github.com/FPcounts/FPEM}
}
