% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_one_country.R
\name{fpem_one_country_autosave}
\alias{fpem_one_country_autosave}
\title{Run one country model and automatically save output}
\usage{
fpem_one_country_autosave(runname = NULL, ...)
}
\description{
Run one country model and automatically save output
}
