% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fpem_plot_country_results}
\alias{fpem_plot_country_results}
\title{plot country results}
\usage{
fpem_plot_country_results(
  country_results,
  observations = NULL,
  first_year,
  last_year,
  indicators
)
}
\arguments{
\item{country_results}{\emph{'Data.frame'} Results data from \code{\link{fpem_calculate_results}}}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}

\item{last_year}{\emph{\sQuote{Numeric}} The last year of model estimates in output. The model will be fit to all data, including dates after this date if available.}

\item{indicators}{name of indicators from results to be plotted}

\item{core_data}{data list from \code{\link{core_data}}}
}
\value{
list of plots
}
\description{
plot country results
}
