% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_plots.R
\name{fpem_plots}
\alias{fpem_plots}
\title{plot country results}
\usage{
fpem_plots(runlist, results, indicators, compare_to_global = FALSE)
}
\arguments{
\item{results}{\emph{'Data.frame'} Results data from \code{\link{fpem_calculate_results}}}

\item{indicators}{name of indicators from results to be plotted}

\item{compare_to_global}{logical, if TRUE plots estimates from global model with dotted lines}

\item{core_data}{data list from \code{\link{core_data}}}
}
\value{
list of plots
}
\description{
plot country results
}
