% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpem_results.R
\name{fpem_results_autosave}
\alias{fpem_results_autosave}
\title{Get results from samples}
\usage{
fpem_results_autosave(runname = NULL, country_population_counts = NULL)
}
\arguments{
\item{country_population_counts}{\emph{'Numeric Vector'} A vector of population counts selected from \code{\link[get_population_counts]{get_population_counts}}}
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
Returns point estimates from posterior samples in long format.
}
