% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpet_calculate_results.R
\name{fpet_calculate_results}
\alias{fpet_calculate_results}
\title{Get results from samples}
\usage{
fpet_calculate_results(fitlist, country_population_counts = NULL)
}
\arguments{
\item{country_population_counts}{\emph{'Numeric Vector'} A vector of population counts selected from \code{\link[get_population_counts]{get_population_counts}}}
}
\value{
\emph{'Data.frame'} A data.frame of point estimates in long format.

\emph{'Data.frame'} A data.frame of point estimates in long format.

\emph{'Data.frame'} A data.frame of point estimates in long format.
}
\description{
Maps multiple sets of runs to \code{\link{fpem_calculate_results}}
}
