% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service_stats.R
\name{list_service_stats}
\alias{list_service_stats}
\title{list_service_stats}
\usage{
list_service_stats(
  service_stats = FALSE,
  service_stats_filepath,
  model_seq_years,
  division_numeric_code,
  first_year = NULL
)
}
\arguments{
\item{service_stats}{\emph{\sQuote{Logical}} If FALSE service statistics are not used. Service stats are not public. Requires FPcounts/fpemservicestat package.}

\item{service_stats_filepath}{File path to service stat csv}

\item{model_seq_years}{Years of model estimation from\code{\link{year_sequence_maker}}.}

\item{division_numeric_code}{\emph{\sQuote{Numeric}} A number associated with the country. See the data from \code{\link{divisions}}.}

\item{first_year}{\emph{\sQuote{Numeric}} The first year of model estimates in output. The model will be fit to all data, including dates before this date if available.}
}
\description{
Filter service statistics data
}
