% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_and_aggregate.R
\name{weight_division_match}
\alias{weight_division_match}
\title{weight_division_match}
\usage{
weight_division_match(division_level_data, population_data, posterior_samples)
}
\arguments{
\item{division_level_data}{data frame with country codes and corresponding aggregate level}

\item{population_data}{population data, union and years of the samples}

\item{posterior_samples}{posteriors samples from \code{\link{posterior_samples_array_format}}}
}
\value{
posterior_samples array with divisions in the order to match weights which will be calculated
}
\description{
weight_division_match
}
